+++ src/textual/_compositor.py
@@ -923,11 +923,8 @@ class Compositor:
         x -= region.x + gutter_left
         y -= region.y + gutter_right
 
-        if y < 0:
-            return None, None
-
-        if x < 0:
-            return widget, Offset(0, y)
+        if x < 0 or y < 0:
+            return widget, Offset(max(0, x), max(0, y))
 
         visible_screen_stack.set(widget.app._background_screens)
         line = widget.render_line(y)